tests/wpt/tests/css/css-fonts/parsing/font-optical-sizing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-optical-sizing with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#propdef-font-optical-sizing">
<meta name="assert" content="font-optical-sizing supports the full grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-optical-sizing', 'auto');
test_valid_value('font-optical-sizing', 'none');
</script>
</body>
</html>
